returns empty dict if no yaml
